Colocate selectors next to gutenberg-gse-settings reducer definition
