+++ client/state/gutenberg-fse-settings/selectors/index.js
@@ -1,0 +1,2 @@
+export { isCoreFSEActive } from './is-core-fse-active';
+export { isCoreFSEEligible } from './is-core-fse-eligible';
