+++ client/state/gutenberg-fse-settings/selectors/is-core-fse-active.js
@@ -1,11 +1,13 @@
+import '../init';
+
 import { get } from 'lodash';
 import shouldCalypsoifyJetpack from 'calypso/state/selectors/should-calypsoify-jetpack';
 
-export const isCoreFSEEligible = ( state, siteId ) => {
+export const isCoreFSEActive = ( state, siteId ) => {
 	return (
 		shouldCalypsoifyJetpack( state, siteId ) &&
-		get( state, [ 'gutenbergFSESettings', siteId, 'isCoreFSEEligible' ], false )
+		get( state, [ 'gutenbergFSESettings', siteId, 'isCoreFSEActive' ], false )
 	);
 };
 
-export default isCoreFSEEligible;
+export default isCoreFSEActive;
